+++ src/main/java/com/atlassian/db/replica/api/AuroraMultiReplicaConsistency.java
@@ -1,6 +1,7 @@
 package com.atlassian.db.replica.api;
 
 import com.atlassian.db.replica.api.exception.ConnectionCouldNotBeClosedException;
+import com.atlassian.db.replica.internal.LazyReference;
 import com.atlassian.db.replica.internal.NoCacheSuppliedCache;
 import com.atlassian.db.replica.internal.NotLoggingLogger;
 import com.atlassian.db.replica.internal.aurora.AuroraClusterDiscovery;
@@ -19,16 +20,19 @@ public final class AuroraMultiReplicaConsistency implements ReplicaConsistency {
     private final Logger logger;
     private final ReplicaConsistency replicaConsistency;
     private final AuroraClusterDiscovery cluster;
+    private final boolean compatibleWithPreviousVersion;
 
     private AuroraMultiReplicaConsistency(
         Logger logger,
         ReplicaConsistency replicaConsistency,
         ReplicaConnectionPerUrlProvider replicaConnectionPerUrlProvider,
         SuppliedCache<Collection<Database>> discoveredReplicasCache,
-        String clusterUri
+        String clusterUri,
+        boolean compatibleWithPreviousVersion
     ) {
         this.logger = logger;
         this.replicaConsistency = replicaConsistency;
+        this.compatibleWithPreviousVersion = compatibleWithPreviousVersion;
         this.cluster = AuroraClusterDiscovery.builder()
             .replicaConnectionPerUrlProvider(replicaConnectionPerUrlProvider)
             .discoveredReplicasCache(discoveredReplicasCache)
@@ -47,6 +51,10 @@ public final class AuroraMultiReplicaConsistency implements ReplicaConsistency {
 
     @Override
     public boolean isConsistent(Supplier<Connection> replicaSupplier) {
+        return compatibleWithPreviousVersion ? isConsistent_old(replicaSupplier) : isConsistent_new(replicaSupplier);
+    }
+
+    private boolean isConsistent_old(Supplier<Connection> replicaSupplier) {
         return cluster.getReplicas(replicaSupplier)
             .stream()
             .allMatch(replica -> {
@@ -61,12 +69,66 @@ public final class AuroraMultiReplicaConsistency implements ReplicaConsistency {
             });
     }
 
+    private boolean isConsistent_new(Supplier<Connection> replicaSupplier) {
+        return cluster.getReplicas(replicaSupplier)
+            .stream()
+            .allMatch(replica -> {
+                try (LazyConnectionSupplier connectionSupplier = new LazyConnectionSupplier(replica.getConnectionSupplier())) {
+                    return replicaConsistency.isConsistent(connectionSupplier);
+                } catch (ReadReplicaConnectionCreationException exception) {
+                    logger.warn("ReadReplicaConnectionCreationException occurred during consistency checking. It is likely that replica is the process of scaling, replica id: " + replica.getId(), exception);
+                    return true;
+                } catch (SQLException exception) {
+                    throw new ConnectionCouldNotBeClosedException(exception);
+                }
+            });
+    }
+
+    private static class LazyConnectionSupplier implements Supplier<Connection>, AutoCloseable {
+        final LazyReference<Connection> connectionLazyReference = new LazyReference<Connection>() {
+            @Override
+            protected Connection create() {
+                return supplier.get();
+            }
+        };
+        private final Supplier<Connection> supplier;
+
+        private LazyConnectionSupplier(Supplier<Connection> supplier) {
+            this.supplier = supplier;
+        }
+
+        /**
+         * The initial implementation of AuroraMultiReplicaConsistency was buggy. The implementation took the
+         * connection management responsibility from ReplicaConsistency. The current implementation fixes the issue
+         * by allowing ReplicaConsistency implementation to take full responsibility for the connection lifecycle.
+         *
+         * Unfortunately, the existing implementations of ReplicaConsistency may rely on the bug (for example some
+         * implementations in our product rely on it). To make it backwards compatible (behavioural compatibility),
+         * LazyConnectionSupplier closes the connection if ReplicaConsistency doesn't do that.
+         */
+        @Override
+        public void close() throws SQLException {
+            if (connectionLazyReference.isInitialized()) {
+                final Connection connection = connectionLazyReference.get();
+                if (!connection.isClosed()) {
+                    connection.close();
+                }
+            }
+        }
+
+        @Override
+        public Connection get() {
+            return connectionLazyReference.get();
+        }
+    }
+
     public static final class Builder {
         private ReplicaConsistency replicaConsistency;
         private ReplicaConnectionPerUrlProvider replicaConnectionPerUrlProvider;
         private SuppliedCache<Collection<Database>> discoveredReplicasCache = new NoCacheSuppliedCache<>();
         private Logger logger = new NotLoggingLogger();
         private String clusterUri;
+        private boolean compatibleWithPreviousVersion = false;
 
         public Builder replicaConsistency(ReplicaConsistency replicaConsistency) {
             this.replicaConsistency = replicaConsistency;
@@ -85,6 +147,18 @@ public final class AuroraMultiReplicaConsistency implements ReplicaConsistency {
 
         public Builder clusterUri(String clusterUri) {
             this.clusterUri = clusterUri;
+            return this;
+        }
+
+        /**
+         * Puts this connection in compatibility mode with a previous version. Developers can use this method to
+         * roll out the new version of the library with a feature flag.
+         * <p>
+         * It's best-effort, and there's no guarantee the library in compatibility mode will always behave
+         * the same way as the previous version of the library.
+         */
+        public Builder compatibleWithPreviousVersion() {
+            this.compatibleWithPreviousVersion = true;
             return this;
         }
 
@@ -120,7 +194,8 @@ public final class AuroraMultiReplicaConsistency implements ReplicaConsistency {
                 replicaConsistency,
                 replicaConnectionPerUrlProvider,
                 discoveredReplicasCache,
-                clusterUri
+                clusterUri,
+                compatibleWithPreviousVersion
             );
         }
     }