+++ CHANGELOG.md
@@ -24,6 +24,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-2.6.0...master
 
+### Fixed
+- AuroraMultiReplicaConsistency#isConsistent always calls replica connection supplier
+
 ## [2.6.0] - 2022-01-26
 [2.6.0]: https://github.com/atlassian-labs/db-replica/compare/release-2.5.6...release-2.6.0